+++ yash-cli/CHANGELOG.md
@@ -25,8 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   or `[[` reserved word is used, indicating that these syntaxes are not yet
   supported.
 - The shell now displays a more informative error message when process
-  redirections (`>(...)` or `<(...)`) are used, indicating that this syntax is
-  not yet supported.
+  redirections (`>(...)` or `<(...)`), pipe redirections (`>>|`), or
+  here-strings (`<<<`) are used, indicating that these redirections are not yet
+  supported.
 
 ## [0.4.5] - 2025-09-20