+++ yash-cli/CHANGELOG.md
@@ -9,7 +9,9 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.6] - Unreleased
+## [3.0.0] - 2025-09-23
+
+This is the first stable release of yash-rs.
 
 ### Added
 
@@ -270,7 +272,7 @@ later.
 
 - Initial release of the shell
 
-[0.4.6]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.4.6
+[3.0.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.0
 [0.4.5]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.4.5
 [0.4.4]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.4.4
 [0.4.3]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.4.3